begin parsing + warning: segfaults for now

+++ srcs/get_next_line.c
@@ -2,7 +2,7 @@
 
 static int find_char(char *str, char c)
 {
-	int	i;
+	int i;
 
 	i = 0;
 	while (str[i])
@@ -14,10 +14,10 @@ static int find_char(char *str, char c)
 	return (-1);
 }
 
-static char	*ft_concat(char *s1, char *s2, int len)
+static char *ft_concat(char *s1, char *s2, size_t len)
 {
-	char	*concat;
-	int		i;
+	char *concat;
+	size_t i;
 
 	if (len < ft_strlen(s2))
 		concat = ft_calloc(ft_strlen(s1) + len + 1, 1);
@@ -33,19 +33,19 @@ static char	*ft_concat(char *s1, char *s2, int len)
 	return (concat);
 }
 
-static void	get_read(char *buffer, int fd, int *read_len, int *cursor)
+static void get_read(char *buffer, int fd, int *read_len, int *cursor)
 {
 	ft_memset(buffer, 0, BUFFER_SIZE);
 	*read_len = read(fd, buffer, BUFFER_SIZE);
 	*cursor = 0;
 }
 
-char	*get_next_line(int fd)
+char *get_next_line(int fd)
 {
-	static char	buf[BUFFER_SIZE + 1] = {0};
-	static int	cs = 0;
-	static int	read_len = 0;
-	char		*output;
+	static char buf[BUFFER_SIZE + 1] = {0};
+	static int cs = 0;
+	static int read_len = 0;
+	char *output;
 
 	if ((!cs && !read_len) || cs == read_len)
 		get_read(buf, fd, &read_len, &cs);
@@ -63,9 +63,8 @@ char	*get_next_line(int fd)
 		{
 			output = ft_concat(output, buf + cs, 1 + find_char(buf + cs, '\n'));
 			cs += 1 + find_char(buf + cs, '\n');
-			break ;
+			break;
 		}
 	}
 	return (output);
 }
-
